Fix for inline SVG icons in IE11
This will resolves #4090 the problem with inline SVG icons in IE11. (The image size must however be set to two integer values and can't handle fractional values, but that is in a compleatly other place in the code.)

Since inline SVG never is loaded determineTainting_() was never called for them, and there is what I have found no way to set that an image is tainting the canvas when you define your image. This code simply makes a tainting check for all images that are loaded when initialized so they get the right tainting value. 

The other part of my fix is to move a line of code into the try statement since it otherwise will cause an error in IE11.