Useful linestring getLength and polygon getArea methods are added to ol.proj and the measure example is updated to use them.
The new methods give good answers in metres, pixels or other projection units whatever the latitude, and whatever the projection.
Results in metres and metres squared are returned for EPSG:4326 and EPSG:3857 projections.
The new methods have an optional ol.Sphere parameter, which might be used for example, with non terrestrial mapping.
By default, ol.sphere.NORMAL is used which has pretty much the same mean earth radius as WGS84.

The custom getPointResolution function attribute of ol.proj is replaced with a constantPointResolution boolean attribute with a getter and setter in ol.proj.
constantPointResolution should be set true for projections such as UTM or EPSG:27700, which have nominally constant scale throughout their extent,
either at construction time, or afterwards. constantPointResolution will default to false unless the projection's units are pixels or degrees,
in which case it is set true. ol.proj.getPointResolution will return an unchanged resolution for constantScale projections.
Projections with constantPointResolution, but not units of degrees, can use faster, more accurate planar cartesian calculations for length and area
rather than spherical ones.

Documentation warnings are added to the ol.geom getLength and getArea methods which for EPSG:3857 and EPSG:4326 do not return results in metres
or metres squared. Also, a warning is added to ol.geom.circle about the radius not being in metres.

Test for the new ol.proj methods are included.