The custom getPointResolution attribute of ol.proj is replaced with a constantScale attribute with a getter and setter. constantScale should be set true for projections such as UTM or EPSG:27700.constantScale will default to false unless the projection units are pixels, in which case it is set true. ol.proj.getPointResolution will return an unchanged resolution for constantScale projections.
An ol.proj.getLength method is added. It returns the length of a line string in metres for EPSG:3857 and EPSG:4326 and other projections with contantScale set false. For all these projections, the points are first converted to EPSG:4326, if necessary, before the length is calculated using the Haversine method. For other projections with constantScale set true, calculation is done using Cartesian maths for speed and accuracy.