Add target option to attribution logo
This commit adds a target option to the LogoOptions. This makes it possible to left-click on an attribution logo and make it open in a new tab. When this option is not set, the target for the logo's link is not set, and it acts like before.

To implement this, I had to change the logos dict format from `{src: href}` to `{src: [href, target]}`.