Add {ModifySegmentDataType#featureSegments} field
An optional {@link ModifySegmentDataType#featureSegments} field has
been added. It is used to list all the modifyable line segments that
are part of the same feature.

This is currently needed for editing circles.  They are broken into two
segments: center and circumference. If the center is modified, both
segments need to be updated.