Add demo based on 3rd party supercluster library
We added the library in the resources as we didn't found it on a
CDN or the official github repository. We named the sample
'supercluster.html' and used the associated 'supercluster.js'.
The particularity of this example is the use of web workers with
'supercluster-worker.js'.
We added the GeoJSON version of the KML earthquakes data as
'supercluster' needs GeoJSON for input.
We may need to correct eslint comment '// eslint-disable-next-line' as
it was the fastest way to make eslint happy with importScripts and
supercluster declarations in the worker file.