Trigger when nothing has been (de)selected
In the past ol.interaction.Select used to fire a 'select' event even when there were no features selected/deselected.  A condition was added to prevent this event from firing when there was nothing selected/deselected.  I'm proposing a new event, "empty", which fires when nothing was selected/deselected.  Without this I believe an explicit map.on("click") handler is required and managing the interaction with the map click and the select interaction gets messy.