Change overlay positioning to use percentage of mapSize, rather than absolute pixel values.
This resolves overlay position appearing differently in print preview
or priting from the browser. I believe the problem is the rasterization of
the image for printing causes a different effective pixel size for the map,
but no javascript events fire during printing to allow map to be reloaded
or overlays repositioned.

By positioning overlay using a % of the parent, it avoids this problem and
the print preview in the browser shows the correct overlay position.