Allow passing featureOverlay to Select
Currently, the Select Interaction creates its own, internal
featureOverlay for housing selected features. Since this is
an unmanaged layer, it is always rendered above the other
layers on the map. To allow for rendering selected features
in some z-indexed position, we must allow users to specify
a managed layer to house the selected features instead. This
will be facilitated via the featureOverlay option.