Do not manage map when featureOverlay specified
Previously, even when a featureOverlay would be specified
by the user, upon selection the feature would always be
added to the very top of the map, and not in the correct
order as determined by the position of the featureOverlay
layer in the map.

We skip this behavior when a featureOverlay is specified,
but keep things as they were in all other cases.