Adding imageDrawBuffer property to VectorLayer
As discussed in #8473, we have a clipping effect when panning on image 
`VectorLayer`. It's because we draw images with viewport size by 
default.

To remove this clipping effect, I created a new paramter for 
`VectorLayer` to specify a buffer to extend the viewport size drawing.

Please note that this parameter will need to increase `renderBuffer` (if 
smaller) to retrieve buffer area features. Obviously, it has an impact 
on performances as the new drawing zone is bigger than the default one.