Restore the rotation/resolution/center constraint system
Constraints are now applied internally by the View class, and the behaviour of
the constraints (i.e. snap to zoom levels or not) is affected by the view hints.

Also a `resolveConstraints` method is available to animate the view to a rest position,
i.e. after a zoom ended between snap positions.