Update MVT.js
descriptions of slice method of Array: 
@param end The end of the specified portion of the array. This is exclusive of the element at the index 'end'. 
initial value of prevEndIndex and i both are 0 so that slice(0, 0)  return a empty array. it lead to a new issue that when I load a multiplygon geometry, endss array's first item is [] but correct value should be the last coordinate index of first polygon of multiplygon. the description of issue: https://github.com/openlayers/openlayers/issues/10500