Use a new cache key in hopes of fixing test failures
The next step in debugging our periodic rendering test failures is to use a new cache key.  This change also removes the less specific key, so we will no longer get partial cache restores.  If we don't see test failures after this change, we may try adding the v2-dependencies- as a key.