Update gom/flat/segments.js to be less confusing to consumers
Calling the callback with the same array over and over, only with its elements reassigned, led to some tricky bugs to track.  By formulating new arrays in the callback call, we avoid having the same array reference presented to the callback on every iteration, so it can do what it likes with its very own copy.