Added 'in' operator as a valid WebGL expression.
Added the 'in' operator for use in WebGL expressions.

Code was written to follow Mapbox specifications, so the expression syntax is ['in', <keyword>, <input>], where <keyword> can be any type, and <input> is either an array or a string.  The expression returns true if <keyword> is contained within <input>.

Added an example entitled "Filtering features with WebGL 'in' operator" to demonstrate the usage of the new operator.