Declare enums before exporting them
This is necessary for getting TypeScript to emit a distinct type alias for an
object marked as `@enum`. For example, consider an enum `MyEnum`. With an unnamed
default export binding

```
/**
 * @enum {string}
 */
export default {
  FOO: "foo"
};
```

the resulting `.d.ts` file looks something like this:

```
declare namespace _default {
  const FOO: string;
}
export default _default;
```

This means that there is not a separate `MyEnum` type to refer to. By changing the enum
declaration and the default export to different lines, e.g.

```
/**
 * @enum {string}
 */
const MyEnum = {
  FOO: "foo"
};
export default MyEnum;
```

the resulting type declaration will contain a generated type alias:

```
export default MyEnum;
type MyEnum = string;
declare namespace MyEnum {
  const FOO: string
}
```

The type alias is necessary to be able to use `MyEnum` as a type.