Exposing clearLoadedExtents in Vector source
In very large datasets, it might be needed to refetch data based on
certain conditions. Such a condition could be a change in zoom level on
the map, leading to more finegrained data.

Right now, it is not possible to reload data with a loader function
after a zoom change, as the loadedExtentsRtree would see the requested
BBox as already loaded, not forwarding the request to the loader
function.

In order to be able to listen to map events outside the vector source
and refetch data based on custom event conditions, the
clearLoadedExtents function was added, so the user can manually decide
when to reset the already fetched extents.