webgl / add stroke & fill styles in literal style format
Note that all the new properties were taken from the flat style format defined
in style/flat.js.

This includes properties for icon, circle and shape, which are redundant with
the current symbol style object.

To avoid introducing any breaking change (even though this API is still
experimental) the symbol style format is kept as is for now.