send 'loadend' event on layer image load failure
Before this fix the map's 'loadend' event was never fired when an error loading a WMS image occured.

The check whether the map finished loaded happens in Map.renderFrame_. That method gets triggered every time something relevant changes in a layer, e.g. when a WMS image finishes loading. But what was missing was the triggering of the rendering when the loading of a WMS image fails.