Don't overwrite browser zoom shortcuts
Some browsers enable the user to set the zoomlevel of the website using `ctrl`
and `+` or `-`. The default version of `KeyboardZoom` catches these keydown
events and adjusts the map zoomlevel.
There might be a situation where a user wants to increase or decrease the size
of e.g. the controls as well. They are not able to do this via shortcuts
because they are handled by OpenLayers. To get around this the default
`KeyboardZoom` should only adjust the zoom level if `+` or `-` was pressed
without `ctrl` being pressed at the same time.  This way, the event is
raised to the browser if `ctrl` was pressed, too.