TileLayerBase: Clip tiles using depth
This changes the WebGL tile renderer to render tiles zoomed-in first,
using the depth buffer as to clip lower layers of tiles without blending
any transparent tiles. Any tile that uses alpha is drawn last, with the
highest possible Z, to allow blending with the lower layer of already drawn
tiles.