webgl / only check winding order of polygons for RenderFeatures
The Feature instances should already come with correctly separated multi-
polygons. The RenderFeature instances can contain multipolygons in the form of
polygons with several outer rings; to handle this we use the inflateEnds
utility which recreates the polygon ends according to the winding order.

Added a rendering test to make sure holes correctly render for Feature instances.