WebGL / correctly compute distance along line when applying an offset
This is necessary for dashes/symbols to render correctly along the line
when an offset is applied.
To do this we need to store an additional information for each vertex:
the sum of the tangents of the join angles encountered so far. This will then
be multiplied by the current line offset in the vertex shader.

Did my best ASCII art to try and explain better what is going on in this part
of the code.