Add coordinates expressions in WebGL tiles
Add the expressions tileCoordX/Y and tileMaxCoordX/Y to tiles using WebGL. This allows testing for the position of the current pixel in style expression. For example, this can be used to handle computations with edges of tiles.
WebGL shaded relief example has been modified accordingly, avoiding artifacts in the rendering.