Add keepUpright parameter
This parameter allows to control whether text can be rendered upside-down
when rendering along a line or polygon. This can be useful when the text
has to follow the line direction regardless of its orientation, like an
arrow pointing the direction of the line element.

This setting already exists in Mapbox Style Spec, so ol-mapbox-style will
be able to follow more closely some style definitions.