Fix non-point geometries in WebGLPoints layer
forEachFeature may skip null-geometries during initialization, this makes the feature
count invalid when it is added to the cache after a change event or when removed.

If there are features with non-point geometries in the cache when rebuilding
buffers, then the render instructions are too long and an icon is drawn in
the map center.

With this commit only features with a point geometry will be inside the cache,
circumventing both problems.