Add option to change TileDebug color
in the case where there are multiple references used at the same time (i.e. with COG), having the same color for the TileDebug is not ideal for readability.

With this new options, each TileDebug instance can now have a different color, giving more flexibility in the case mentioned above.