Expose anchorOrigin and anchor[XY]Units from ol/style/Icon
While migrating my code to TypeScript, I encountered a problem because I was using `anchorOrigin`, `anchorXUnits` and `anchorYUnits` to better decide what to do with my application icons.

As they are labeled as `@private` in the JSDoc, they are automatically made private members of the Icon type definition in TypeScript and can't be accessed.

So I propose we expose them, through a function, so that they may also be used/read in a TypeScript context.