VectorLayer: Include source extent in maxExtent
When rendering multiworld, the builder/executor need to be able to
draw features that fall into a wrapped version of the view extent.

To accomplish this, this change watches for changes in the source
extent (so that we don't have to do all this every frame), re-
projects it into the view extent if possible, and then extends
the extent to be used for maxExtent.