VectorLayer: Use source extent for maxExtent
When rendering multiworld, the builder/executor need to be able to
draw features that fall into a wrapped version of the view extent.

To accomplish this, this change watches for changes in the source
extent (so that we don't have to buffer every frame) and uses
the source extent for maxExtent when available.