VectorLayer: Ensure all wrapped features get rendered
Some features extending beyond the projection extent need one
(or more) extra worlds to be rendered on either side. This can
calculated by comparing the farthest reaching feature on one side
of the source to the opposite bound of the extent.

Put another way:

The left-most center for a world that needs to be rendered, is the
left-bound of the viewport minus the right-most feature from the
source. Divide by worldWidth and round.
Vice-versa for the right-most.